requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
python-multipart
jinja2
redis
fuzzywuzzy
python-levenshtein